requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
rank-bm25
requests
pydantic
jinja2
python-multipart
pdfplumber